[FE: REFACTOR] 코드 스타일을 Vue2 -> Vue3식으로 리팩토링
