+++ f_end/src/components/login/GoogleButton.vue
@@ -1,22 +1,13 @@
 <template>
 <div style="display: flex; flex-direction: column; align-items: center">
   <img
-    src="@/assets/kakao_login_large_narrow.png"
+    src="@/assets/btn_google_signin_light_normal_web@2x.png"
     height="92"
     width="382"
   />
 </div>
 </template>
 <script>
-export default {
-  name: 'KakaoButton',
-  data() {
-    return {
-    }
-  },
-  methods: {
-  }
-};
 </script>
 
 <style scoped>
